\docType{data}
\name{mESCs}
\title{100 mESCs example data}
\alias{mESCs}
\format{A dataframe of 3134 rows (genes) and 100 columns (cells) of CPM normalized read counts}
\usage{mESCs}
\description{A subset of the Kolodziejczk et al. (2015) mESC data. 50 cells from the a2i and 50 cells from the lif growth conditions were randomly sampled. Genes were filtered to remove lowly expressed and genes (mean expression < 5) or without correlations of magnitude >0.2 with other genes.
}
\value{A dataframe of 3134 rows (genes) and 100 columns (cells) of CPM normalized read counts}
\source{\url{https://www.ebi.ac.uk/teichmann-srv/espresso/}}
\keyword{datasets}

 
